task/requirements.txt
Django==2.2.18
djangorestframework==3.12.4
django-cors-headers==3.7.0
psycopg2==2.9.1
gunicorn==20.1.0